CS 204 – Advanced Programming-  HW1
The aim of this homework is to practice on the topic of two dimensional (2D) matrices implemented using vector of vector. I useD a vector of vector of char (i.e., char matrix) as the data structure to store and manipulate the data. The matrix content will basically be characters representing four main directions, and a character representing empty cell. The initial content of the matrix will be read from a text file. ı will also read integer input from the standard input (keyboard). According to the input to be taken from the keyboard and the values on the matrix, code will move the tiles in the cells toward its direction until it is blocked by another non empty cell or drops off from the edge of the matrix. Additional information is given in the rest of document.